Removed use of slices.Equal for Go 1.20; will use it for Go 1.21 and above
